minor adjustment for bool conversion; change targets back to float

+++ selene_sdk/train_model.py
@@ -534,7 +534,7 @@ class TrainModel(object):
         while count < data_targets.shape[0]:
             remainder = min(data_targets.shape[0] - count, self.batch_size)
             inputs = data_seqs[count:count + remainder, :, :]
-            targets = data_targets[count:count + remainder, :].astype(int)
+            targets = data_targets[count:count + remainder, :].astype(float)
             inputs = torch.Tensor(inputs)
             targets = torch.Tensor(targets)
